pyproject.toml
[project]
name = "morning_brief"
version = "0.1.0"
description = "Morning consumer trend briefing CLI"
requires-python = ">=3.11"
dependencies = [
    "feedparser>=6.0",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "pydantic>=2.6",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "anthropic>=0.40",
    "python-dotenv>=1.0",
    "rapidfuzz>=3.6",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-mock",
]

[project.scripts]
morning-brief = "morning_brief.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["morning_brief"]
